[MAJOR IMP] Rewrite the http/RPC engine and let HTTP/1.1 features.
This patch attempts a major change in the structure of the XML-RPC
framework. There is one http server, capable of multiple services
(except XML-RPC). That server could handle authentication, and is also
HTTP/1.1 capable, which means it supports **persistent connections** !

At this commit, the old behaviour of the XML-RPC protocol is merely
working. The netsvc.Service is split, so expect wizard/report breakages.
External modules (koo) also break with this API.

The net-svc code is crippled and gone FTM.