[IMP] Base_module_record: export changes as xml
Instead of a complete zip file, export the selected data changes as
a xml text segment. That is for developers, that want to edit stuff in
a database and transfer those changes back into their xml data files.

Also, clean the format of text nodes in the xml. Make it more readable.