convert: When loading menu, don't match by name
It is safer to create duplicate menus than assume the same name is the same
menu.

Solved strange case with 'Manufacturing', where it would falsely identify
the 'Manufacturing/Manufacturing' as its parent, and thus create a recursive
menu. If we are parsing the 'a/b/c' name syntax, we have to ommit "c" from
parent search.

Conflicts:

	bin/tools/convert.py