account: fix cancelling of invoices, after e27a55bc59d
In that commit [Bug 675494], the account.move was given a
constraint not to unlink records referenced by an invoice. This,
however, created a logical deadlock with invoices that tried to
cancel and unlink their moves.

Now, change the order so that the move is first detached from
invoice, then unlinked, and then have the invoice be cancelled.