[WIP] all: (work in progress) changing wizards to avoid using read() or at least take proper care of m2o fields
This is in preparation of the fix for m2o.get() on osv_memory objects, to make
it consistent with the regular read on osv objects. Previously on osv_memory
objects it was only returned the id, instead of the pair (id, name) as is appropriate
to avoid an additional name_get RPC call after each read.
Unfortunately many wizards were using direct read() instead of browse() and
relying on the fact that the returned dict contained a bare id value for
many2one fields.