[FIX] share: use empty domain for shared actions, as ir.rules do enforce the filtering.
We must not use a normal domain for the shared action,
because it may be constructed with records the user may
not see (e.g. sharing tasks from a certain project does
not give right to see the project itself.)
Evaluating the domain of an action requires rights that
the user will not have. Moreover, this is useless because
ir.rules are created with each share access to filter
appropriately the records, and they are evaluated as 
root user without risk of failed access right.