[FIX] add wrapper element around kanban group titles
Behavior of position: relative is undefined on all table elements,
Webkit choses to interpret it whereas Firefox decides to ignore it.

There is a position:relative on the group title (to position the title
text), and because it's ignored in Firefox the titles position
themselves relative to the application frame itself. By adding a
wrapper element within the table cell and moving the relevant class
onto it, the position:relative becomes "correct" (specified behavior)
and taken in account by everybody, positioning the title correctly in
Gecko-base browsers.