[FIX] move reload-webclient-on-module-install to act_window execution
A special case reloads the Webclient if a dialog-included operation
involving potential module installation (base.module.upgrade, or
base.setup.installer) is involved.
      
This was performed on an act_close, but it meant a module-involving
act_window followed by a non-module-involving act_window
(e.g. installation of a chart of account) did not match the check
criteria and failed to trigger the reloading of the web client.
      
Check moved to act_window execution, and augments the on_close handler
bound to the dialog's closing any time a potential module installation
(see above) is encountered, to ensure we don't "lose" cases, such as
the installation of the `sales` application (from "add more
features"), which was the issue in the related lp bug.