[IMP] document_ftp: improve/sanitize transaction management during tests
Now that spurious commits were removed in document module,
the test would not run properly anymore.
We do need special care with transactions when we test the
FTP layer because operations on the virtual FS are done
in their own transaction, and immediately committed.
In order for test code and FTP operations to be able
to see each other, the transactions need to be synchronized. 
The current test transaction needs to be rolled back
or committed after reach FTP operation so it can "see" the
latest changes in the database. Similarly, we must force
a commit of the test transaction whenever we want the
FTP layer to be able to see its effects.