[FIX] race condition on selecting saved filters
Deferreds can't "go through" al callbacks, so even when searchviews
correctly wait on this.on_clear it can't get the information on when
the other view is done clearing, and it lauches two concurrent views.

In this precise case though, there's no need to launch a search after
this clear, just want to remove all searchview state.