[IMP] orm: support model-level @groups attribute for access restriction
This is a partial implementation with no support for
restricting read/write access via RPC. This first
part only covers the removal of the restricted fields
from the client-side view, i.e. in the results of
fields_view_get() and fields_get().

The second part will come later and will cover the
real low-level access control.