[FIX] data_import: allow importing m2o via name_search() on top of their XML ID
We do default to exporting the XML ID, but there are
many cases where completely new data needs to be
imported and there is no XML ID available to do so.
In that case allowing name_search() to be used as in
6.0 is a life-saver. This patch simply makes the
m2o field itself visible during import, so users can
use it in their CSV file or manually select it.

Some trivial examples:
- you need to import 2000 new leads with their countries,
  obviously you can't guess that base.be should be used
  for country Belgium
- you need to update 2000 unassigned leads in the system
  and assign them to some of the salemen... but you don't
  have their XML IDs and even if you had, the match is
  quite hard to do.