[REVERT] db: revert incorrect change to auto_install logic
The auto_install flag means that the module will be automatically
installed as soon as all its dependencies are satisfied.
It does *not* mean that the module will be automatically installed
upon database creation. It can be used for that purpose by
setting it on a module that has no dependencies however.