[FIX] Create new records as a bunch of empty (false) fields, otherwise Bad Things Happen                                                                                        ~/projects/tiny/web/current
namely, if the list view fields have e.g. attributes associated, the
computation of the domains blow up

also, always create an editor in the listview (if the editable
listview module has been installed), avoids blowing up 'safety' calls
to #ensureSaved. An alternative would be to fix #ensureSaved not to
blow up if there's no editor, but that means third parties which
*know* there may be an editor in the list view can't easily hook up to
it.

Things will have to change anyway as currently toggling a list view
from not-editable to editable after on_loaded has been called will not
work correctly (it won't start the editor), which is shitty.