[IMP] validation of search_default values for m2o fields
Just because the search_default is an array does not mean it's a
name_get apparently: there are many cases of a single id wrapped in an
array (``[id]``).

So ensure that the search_default is an array *of length 2 whose
second element is a string* before considering it a name_get and using
it as-is (without resolving the label via an explicit name_get).

Also, error out for any m2o search_default which is an array of length
> 1 (unless the second element is a string as per description above).