[ADD] reload file button next to the file input, as webkit browsers and MSIE don't trigger change when a file input is re-set to thee same file
Using other events (e.g. 'click') is not kosher as the event is
triggered long before the file is picked, the alternative is to use
the File API and regularly check the file in the input to see if it's
been altered (by storing a crc) and re-pushing it if the hash is
different. Of course this has other issues (risks catching the file
mid-edition...)

A third alternative is, once a file has been uploaded, to clear and
re-insert the corresponding file input. This means the file input will
*always* be empty