[FIX] reorderable listview clone/placeholder positioning
In some cases (tables?) jquery-ui sortable apparently breaks down if
any parent of the element being positioned (and the generated helper
element) is in position: relative.

Remove all of the position:relative in the way, they seem to not be
used anymore (including the one in jquery-ui, testing in IE9 doesn't
seem to expose any bug and jquery-ui references one, might have been
an issue in IE<=8 which was fixed since).

Maybe an assertion should be added to the code, which would call
#getComputedStyle() on all the #parents() of the element being made
sortable and error out if one of them has
#getPropertyValue('position') == 'relative'?