[FIX] editable listview: do not attempt to delete falsy ids (from the dataset)
During the creation of a new record (in an editable list), the
prospective new record has a ``false`` id. In a non-o2m, the row under
creation can be selected (checkbox) and deleted (More > Delete),
leading to the request to ``DELETE false`` being sent to the ORM…
which said ORM does not wish to entertain.

Before actually performing the delete request (more precisely
forwarding said request to the listview itself), remove all false-ish
ids from the array.

Also, don't send a request to delete nothing.