[FIX] Modified behavior when we click on an action in the 'More' menu.
Since revision 1517, when using the sidebar from a form view, it merged all the fields of the form view in the context of the action.
It was modified to avoid that context pollution, now it only uses those fields to evaluate the context of that action. So it allows
specifying contexts like "{'invoice_id': invoice_id}", but invoice_id will not be propagated any more to the new action context if not
explicitely specified.