[]FIX] disable (broken) inline edition in grouped editable lists
Because the Add/Create button is global, it's not really possible to
know *where* to create the new record, and the current design can't
really deal with a record being edited outside of all groups.

An option might be to create a dedicated empty group for that, but the
cost/benefit is high and the UI would probably be odd.