[FIX] visible text in Chrome on Linux after applying text-to-icon
@include text-to-icon makes the text invisible (and adds an icon), but
some platform apparently don't apply the text's transparency (or
visibility) to its shadow, leading to the invisible text's shadow
still being noticeably visible (an interesting literary idea, an
"invisible" man who still casts a shadow?) as if the text had not been
erased. Furthermore, combined with the 1px font size it made the text
look like some sort of screen smudge or glitch.

Fix by erasing the text-shadow as well as the text.