[IMP] account: Make partner_id consistent in invoices
Both account.invoice and account.invoice.line contains a reference to the
partner_id. The partner_id field in the line is a related one, so it should
be automatically computed, but we have known through issues/not accepted
patches like https://github.com/odoo/odoo/pull/6774, that this doesn't
work correctly.

This desynchronization happens only on certain models, and
account.invoice.line is one of them. If the old linked record is removed,
there's also no restrict constraint, so we find a foreign key error
when updating.

This commit synchronize again all desynchronized records.