[IMP] mail migrations: Avoid inserting orphan followers.
Since the mail_followers table does not have (can't have) a foreign key
restriction, some followers may be subscribed to long gone mail_groups.

This kind of error does not show up in the application and may passed
undetected to users.

However, when inserting the followers into channels the query fails with a
non-existent channel.  This commit avoids the error and, nevertheless, there
won't be a loss of information from the application's users.