[FIX] *: Handle attachment=True property
This commit patches all the existing migrated modules for fixing
the binaries on fields with attachment=True property set in v9.

NOTE: There's still some modules pending, like fleet or gamification,
with this kind of fields, but as there's not yet migration script for
them, it's better to not add anything for not creating false sensation
that the module is migrated.