[IMP][9.0] account: Copy the currency_id to invoice lines.
The currency_id field of an invoice line is a related field taken from the
parent invoice.  However, it is computed the same way all other functional
fields are computed and it may take a long time if you happen to have many
many invoice lines.

Solution: blacklist the currency_id of invoice lines and fill it with SQL.