[IMP] account: Merge invoice journals
In version 8, Odoo had 2 journal types for each invoice operation
(sale and sale_refund, purchase and purchase_refund). In v9, now both
are merged in one journal, and this eases the view of all the operations
together.

If we use a migrated v8 DB, we still see both journals, but creating
new refunds are made on the same journal than the normal operation,
splitting the invoices in 2 journals.

With this, we move the old refunds to the normal journal if it can be
determined (if there's only one journal).

For cases where the journal layout is more complex, you can pass
arguments for mapping the journals, although it has to be manually
hacked on each installation, or look for another automation.