[FIX] compare_noupdate_xml_records: Don't ignore forcecreate
Records with forcecreate="0"/"False" are not recreated when they
have been previously removed. If we try to load some changes in
records of this type and the records have been removed, we will get
for sure an error of missing required fields.

Preserving the attribute on changed records the check for existence
is done and you won't get errors