[IMP] account: Merge supplier invoice references
In v8, there are 2 fields for writing references:
supplier_invoice_number and reference. Now in v9 there's only the last one.
We merge the first field content in the second one for avoiding data loss.
Note that previously the `reference` field has been copied for preserving
the original field contents.