[IMP] account_voucher: Use the voucher's number before comment.
In a DB migration, I've found that most of the payments end-up being named
'Write-off' which is the default comment in Odoo 8.  This is because our
clients use preferably the number instead of the name.  The comment is
normally not changed because it's set when there are differences between
credits and debits which is rare in our case.

Proposal: Use the number instead of the comment first.