[FIX] Do not drop indexes
Odoo will drop any index that is defined in any other module than the module
that adds the column itself. Such indexes were supposedly added by customization
modules because they were needed in real life scenarios. Typically, removing
such indexes are harmful to the performance of the migration, plus it adds the
cost of re-adding the indexes on the upgrade of the module that defines them.