[FIX] stock: Use pack operation > move link qty
When having several stock.move with the same product, Odoo grouped them on the same
stock.pack.operation record, so use it for determining the quantities on stock.move.line
records (at least one per stock.move), will make this quantity to be incorrectly higher.

We also discard stock.pack.operation.lot information, as this doubles the records
on the query, and we can recreate that information using the lot of the quant linked
in the stock.move.operation.link record.

Tried this query in:

- Incoming transfer without lots
- Incoming transfer with lots
- Outgoing/Internal transfer without lots
- Outgoing/Internal transfer with lots

Finally, we only insert move lines for done moves, as for those not done, there's
no need of creating stock.move.line, as they can be recreated in v11 easily:

* For outgoing or internal transfers, clicking on "Check availability", which
  takes current reserved quants information for recreating the information
* For incoming transfers, clicking on validate for accepting full quantities,
  or writing the quantity in the lines you want in "Operations" tab. The only
  drawback is that if you want to use the detailed operation mode, you will
  have to enter quantity manually.