[FIX] tax migration algorithm, simplify more
It appears that attempting to reuse existing taxes is
a bad idea. Indeed, when running account_chart_update after
migration, tags will be overridden possibly changing past VAT
declarations. So we simply create one inactive tax for each tax tag
we need.

Moreover we provide guidance to determine if a positive value
in tax_amount must be a debit or credit.