[FIX] stock: Don't fail on incompatible conversion
After stock.move.line creation, all records are checked to see if the have different
UoM from product one, and re-compute `product_qty`. If you have changed improperly
the unit of measure of the product to an incompatible UoM (this was possible to
be made on <= v8, or now deactivating the product, changing the UoM, and activating
it again), the conversion will fail.

We enclose it in a try for ignoring these conversions.