[MIG] website: Handle Bootstrap 3->4 and Less->Scss
In case a user has customized any Less attachments, these are disabled because 99% of the times it will fail: Syntax from Scss and Less is different, there may not be a `lessc` executable in the Odoo 12 environment, and inherited views would probably fail finding the node to replace from XPath.

Automated resilient conversion between Less and Scss is close to impossible.

To avoid problems, these views are disabled. You are still able to fix and reenable them manually as needed.

In Odoo v12, the standard behavior when a user modifies a website template is to create a modified copy which is specific to the current website (a.k.a. COW 🐮).

To benefit from this new behavior, when migrating a website from Odoo 11 to 12:

1. If a website-specific view exists, it is migrated in place.
2. Website-agnostic views are migrated in a COW 🐮 view, specific for each preexisting website.
3. Webiste-agnostic views are marked as `noupdate=0`, so when starting a new website, you have the updated views as a starting point. To make this work, all this conversion had to be made in the pre-migration stage.
4. The new copies will have a hidden migration column that indicates from which view was it copied.
5. If the views were using any default images from snippets, they will be stored in the filestore now.

To make this work, https://github.com/OCA/openupgradelib/pull/148 is needed.