[FIX] stock: Mark references as noupdate=1
Main warehouse references records have an associated XML-ID, that was
updated in v11 through a method and maintain as XML noupdate=0 data, so
they weren't removed on updates, but now on v12, that XML-IDs are
noupdate=1, and no XML data in provided, so on regular update process, they
are tried to be removed. We switch them to noupdate=1 for avoiding this problem.

Fixes #1925