[12.0][FIX] base: split v11 admin user into system and admin in the post-migration.
Before this the split was done in the end-migration and any other module modifyting
base.user_root will be modifying the previous v11 admin. After that the split is done
in the end-migration of base and you end up with the modifications meant for the
v12 __system__ user in the v12 admin user.