[12.0][FIX] mail_bot: set auto_install to false.
This module is not adding any core functionality, only an
automatic chat that helps to discover funtionalities. Since it
is a new module in v12 it was always being installed by
openupgrade, this causes some problems with admin user split
done in end-migration of base module (see
https://github.com/OCA/OpenUpgrade/issues/1953).

If after running openupgrade you start you migrated database
with upstream Odoo, mail_bot will be installed but this time
without generating issues in the migrations result.