[FIX] Reverts #706
The change has no base as IDs are not being changed and both product.template and
product.product exists in v7 as well.

This is a bad interpretation of migration with individual needs that has been prioritized
as general. If someone wants to move attachments from one model to the other, he/she
must do it by its own.

Fixes #1912