[IMP] base: Disable VAT validation when syncing partner commercial fields
This method took about 8h in a database with 280000 partners. It's disgusting when all that time is wasted just because one partner failed VAT validation, and it's even more disgusting when you don't know which one failed. When VIES validation is in place, it can take several hours more.

Now, in case a partner cannot sync its commercial fields for any reason, in the log you'll be able to see the whole exception message and the partner ID for each one of those that failed. At least, you'll be able to fix all migration problems at once before retrying the migration.