[IMP] Futher details when failing a partner
The ID isn't enough to know the reason of a partner failure, because in v7 migration many new partners are created. I faced a case where the failing partner had an ID that wasn't existing in the v6.1 database, so the ID didn't help.

I'm logging the name and VAT (which is the field that will produce these failures most times) to make these logs even more helpful.

@Tecnativa TT18838