[FIX] Regression in matched_percentage calculation causing division by zero
Main problem is that BETWEEN is inclusive, which allows for a rounding of 1
as is the case for JPY. The LOG of 1 is 0 which causes a division by zero.
Instead, we can just rely on NULL to invalidate numeric comparisons and check
if the result stands up to the numeric comparison 'smaller than 1'.