[FIX] base: Drop FKs for procurement.order
Running [this v12 migration code][1] in a DB that was migrated from v10 and that had many procurement orders takes days. That's because the FK was set without an index.

At the end of the day, this isn't useful because that table is no longer used, so I'm dropping its FKs here.

WIP until merged:

- [ ] https://github.com/OCA/openupgradelib/pull/195

[1]: https://github.com/OCA/OpenUpgrade/blob/bdf26ceb11834fd45ca4b4577532d6f5ee2ca4e4/addons/sale/migrations/12.0.1.1/pre-migration.py#L93-L104