[FIX] stock: Tracking by serial assignation by product, not by template
Previous query assumes that product template IDs are the same than the product
variants IDs, which is not correct when you have variants.

With this new query, that's taken into account, although there's the risk that
one variant has lots with one quantity and others not, and this query assigns
serial in that case.