[FIX] sale: Avoid costly qty_delivered computation
In v11, `qty_delivered` was a regular field that was filled by business logic.
Now in v12, it's a computed (stored) field, which means that it's automatically
computed according some conditions.

As there's a chance that the stored value differs from the v12 computation result
(for example, if you have extension modules that modifies the logic), we provided
a trick for computing on the fly the value for v12, and if it differs from the
stored one, we switch the delivered method to manual and set the previous value.

This has proven to be extremely slow with large quantity of sales order lines
(which is logic, as it requires to create a virtual record per record, and
compute by ORM the result).

That's why this approach is disabled in this commit, given the facts:

* It's unlikely that the stored quantity differs from the v12 computed quantity.
  If you had v11 modules that alters this result, you should replicate them on v12.
* Even on that case, the stored delivered quantity is preserved while no change
  in dependant fields is triggered, so for most of the records, the value will
  be untouched.