[FIX] base: avoid unique_name_per_day sql constrain
This constrain is new in v11. As a result, in v10 you could have two records with the same name, currency_id, company_id and not violate the constrain, but in v11 the name is converted from datetime to date, and then it fails.

Also fixes tz from first user's company to create_uid user.