[FIX] mail: remove admin from mail.alias
When a mail.alias record belongs to admin, it will try to send mails in name of a user that most likely has wrong data in it, and is actually garbage. That can produce a problem where aliases create records but don't notify followers, or even worse, that same problem in loop (depending possibly on if you use IMAP4 or POP3 to fetch mails).

@Tecnativa TT24347