[FIX] hr_timesheet: widgets for UoMs
This is a corner case that can't be automatically handled by static noupdate compare script:

- This is not a full record definition, but updating one field.
- The original record definition is noupdate.
- This update is not.

This is working when installing the module, as Odoo updates noupdate records in init mode
when they are noupdate in its redefinition, but it's not suitable for updates/migrations.

Fixed putting manually on the noupdate XML to load both records.