[FIX] account: Don't populate invoice lines on not subjected VAT
There's a tax in Spanish localization of "VAT not subjected", that
creates this journal items:

1- Expense account: Base amount
2- Expense account: Tax amount
3- AR account: base + tax

In this case, the second line was taken as invoice line, creating a fake
extra move line for it, and also labelling it as "(OLD GROUPED ITEM)".

We avoid both adding the clause of not having a tax_line_id.