[OU-FIX] #2339, set noupdate of ir.model.fields XMLIDs to FALSE
Fixes #2339

Also

* add loaded models to the set of loaded XMLIDs
* purge models with noupdate NULL instead of FALSE
* Remove redundant update_fields_xmlid method and add tests
* fixes broken migration tests mechanism
* fix unreleased savepoint
* improve logging when deletion fails (should be rare now)
* remove coverage as it is giving out red marks to PRs for no reason
* reduce logging of deletions in CI
* remove ir.model.relation from ir.model's unlink

This PR is part of a set of PRs for OpenUpgrade 9.0 up to 13.0. Background:

While Odoo deletes obsolete field and model entries from the data model
metadata explicitely in their migration scripts, in OpenUpgrade I have
always meant to rely on the mechanism of purging 'untouched' XMLIDs that
takes care of the deletion of obsolete data records (e.g. views).

However, this mechanism was not applied to field and model entries because
their XMLIDs were created with noupdate NULL instead of FALSE and as such
excluded in the query to gather all obsolete data records.

Also missing was marking the XMLIDs of fields and models as loaded in the
first place.

All of this is working properly in Odoo 13 (introduced gradually across new
releases) so all of this is backported from newer versions one way or
another.