[OU-FIX] sale: group_auto_done_setting should be enabled by default
In v12, any user could lock the sale order when they wanted (independently if it was done automatically or not by the config setting) because the button action_done was accessible. But now in v13, the button only is accesible with the setting enabled. In order for the button be accessible (as it was in v12), then the group_auto_done_setting should be enabled.