[FIX] product: Insert missing template attribute/value records
Given this situation in v12:

- Template T with attribute A and values A1, and A2, and attribute B with value B1.
- Generated variants V1 and V2 for attribute values A1/B1 and A2/B2 respectively.
- Generated product.template.attribute.line records for T/A and T/B.
- V2 is used in a quotation.
- Remove B attribute from the template. Result:
  * V2 is archived
  * product.template.attribute.line T/B is removed.

On v13, the record is not removed, but marked with active = False.

From the current data status we find on v12 for these cases, we need to
reintroduce missing product.template.attribute.line records with
active = False needed later on next steps for DB integrity.

----

Given this situation in v12:

- Template T with attribute A and values A1, and A2.
- Generated variants V1 and V2 for attribute values A1 and A2 respectively.
- Generated product.template.attribute.value for T/A1 and T/A2.
- V2 is used in a quotation.
- Remove A2 attribute value from the template. Result:
  * V2 is archived
  * product.template.attribute.value T/A2 is removed.

On v13, the record is not removed, but marked with ptav_active = False. That's
because there's a field in product.product called combination_indices that
stores the ID of such line and serves for quick search/indexing on it.

From the current data status we find on v12 for these cases, we need to
reintroduce missing product.template.attribute.value records with
ptav_active = False for not having later unique constraint errors and
proper DB integrity.

This is also the second step for amending the situation described in
``insert_missing_product_template_attribute_line`` method.