[OU-FIX] account: Don't disable payment term multi-company rule
This rule is introduced one version after introducing the `company_id` field in
`account.payment.term`. At this level, we can have payment terms created for one
company, but used in documents of other company, so it doesn't serve to disable
the rule for avoiding the mess, and more as in v13, the company consistency is
checked per document.

Let's keep the rule enabled and get the yell before going to v13 then.