[OU-FIX+IMP] product: Populate all required ptav + fill attribute_line_id
An extra link requirement is added on the LEFT JOIN that checks if the
proper product_template_attribute_value exists. If not, there can be
false matches that prevents the creation of the line.

Steps to reproduce (that partially fixes this commit):

- Create a Product Template with:

  Attribute lines: Color (Red, Blue), Material (Iron, Wood)

- Create and Validate transfer with Variant (Red, Iron)
- Remove Red value from attribute line
  (on this step we lose here link between product.attribute.value and product.template.attribute.value)
- run migration
- Go to Product Template form
- Press on Configure Variants button
 -Set Inactive filter

Current result: Missing attribute for inactive product.template.attribute.value.

We also fill at the same time the attribute_line_id values for this
newly created records, for not having to be filled later.