[OU-FIX] product: fill attribute_id of ptavs in pre-migration
When the module is loading, the new product_attribute_product_template_rel
table will be automatically created and filled (it's an stored computed m2m).
Thus, the attribute_id of ptavs is needed in such compute, and we need
to move the filling of the attribute_id of ptavs to pre-migration.