[FIX] account: Don't have duplicated invoice line links
With previous code, same invoice line may be linked to several
account move lines if there are several of them with same product,
quantity, amount and the rest of the distinct data that differentiates
them, provoking later problems on the rest of things that depend on
this (sales lines linked to the old invoice line for example).

With this iterative algorithm based on a divide-and-conquer strategy,
we fill according that criteria, but excluding already linked invoice
lines, and emptying duplicates from the second appearance onwards. The
worst case will mean at least as many loops as the number of identical
invoice lines we find in one invoice, but the time required for the
SQLs is decreasing, so we can end the loop on an affordable balance
between spent time and results.

Second attempt after reverting due to performance problems #2614. This
one has been inspired on that one, but with several key changes for
performance:

- Don't do a costly extra LEFT JOIN over invoice lines.
- Cache in an extra column if an invoice line has been already linked.
- Use such cached value for excluding invoice lines for next iteration.

Co-Authored-By: mreficent <miquel.raich@eficent.com>