[FIX] use_parent_address <=> partner type 'contact'
The usage of use_parent_address is replaced by address type logic
(see https://github.com/odoo/odoo/commit/43839a84), even if the field
is only removed in Odoo 10.0.
Addresses are now synced with the parent if the partner type is 'contact'.
Therefore, set addresses of type 'other' to 'contact' if they had the
`use_parent_address` flag in 8.0, and vice versa.