[IMP] website: Try to remove googleplus references
On v13, the field `social_googleplus` has been removed, but it was put
on certain places on the website, and they remain unaltered due to the
noupdate=1 flag or being a COW (created on write) view, so we directly
remove that part from the view if we find the exact HTML expected code
that is on the definition. This is done for avoiding side effects, and
it means that if you have altered somehow that part, you will need to
remove it manually.

Fixes #2886