[FIX] delivery: invoice_policy
Odoo changed the behavior of delivery charges from v12 to v13 delegating
the mode check into the new `invoice_policy` field in
`delivery.carrier`.

Initially, it was decided to keep the behavior in v12 that recomputes
the price when the picking is validated, but there's no accurate flow
translation between versions and some issues raise with this default:

- Core UI forbids users changing this value in the standard shipping methods.
  Only extra carriers can choose it and the default `estimated` value
  remains.
- The eCommerce behaves weirdly with "Real cost" methods, leading the
  customers to think that they get free delivery orders.

For that reasons, a sane default must be Estimated cost for the majority
of cases.

TT31975